+++ changes.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Support Python 3 as well.
 
+- Set a timeout during initial requests.
+
 
 1.0 - 28 August 2015
 --------------------